db_server/requirements.txt
requests
requests-futures
psycopg2-binary
pyTelegramBotAPI
beautifulsoup4